Bring all notebooks up to date. (#12)
Re-rendered all notebooks.

Fixed two issues:
1. notebook `R - How to pull in reusable code from source control.ipynb` was installing a particular older version of the [`gargle`](https://github.com/r-lib/gargle) package as an example. BUT [`bigrquery`](https://github.com/r-dbi/bigrquery) now depends on gargle for auth logic. Downgrading gargle causes bigrquery auth to stop working.
2. The input data URL for notebook `R - Plotting example for Boston 311 data.ipynb` no longer worked (returned a 500 level error). I updated the notebook to fetch the data from a newer URL.

Otherwise all the notebooks were working fine as-is. In several cases I removed code that was no longer needed. For example, I removed obsolete calls to the Ronaldo package. With the most recent release of bigrquery https://github.com/r-dbi/bigrquery/releases/tag/v1.2.0 it now uses the gargle R package for auth. This is great because if there is already a credential on the machine, it will make use of that by default. [less friction for the user!]